Private Hosted zone
Welcome back, and in this video I want to talk about the other type of hosted zone available within route 53 And that's private hosted zones. So let's jump in and get started straight away. A private hosted zone is just like a public hosted zone in terms of how it operates. Only it's not public, instead of being public. It's associated with VPC is within AWS, and it's only accessible within V PCs, that it's associated with, You can associate a private hosted zone with three PCs in your account using the console UI COI an API, and even in different accounts, if you use the COI an API only. Everything else is the same, you can use them to create resource records, And these are resolvable within V PCs. It's even possible to use a technique called Split View or split horizon DNS, which is where you have public and private hosted zones of the same name, meaning you can have a different variant of a zone for private users versus public, you might do this if you want your company intranet to run on the same address as your website and have your users be presented with your intranet when internal, but the public website when anyone accesses from outside of your corporate network, or if you wanted certain systems to be accessible via your business's DNS, but only within your environment. Now let's quickly step through how private hosted zones work visually so that you have more of an idea of the end to end architecture. So we start with a private hosted zone and as with public zones we can create records within this zone. Now from the public Internet, our users can do normal DNS queries so for things like netflix.com and cartogram.io, bought the private hosted zone is in accessible from the public Internet. It can be made accessible though, from V PCs, let's assume all three of these V PCs have services inside them and use the route 53 resolver so the V PC plus two address. While any V PCs, which we associate with the private hosted zone will be able to access that zone via resolver. Any V PCs which aren't associated will face the same problem as the user on the public Internet, on the left, access isn't available. So private hosted zones are great when you need to provide records using DNS, but maybe they're sensitive, and need to be accessible only from internal V PCs. Just remember, to be able to access a private hosted zone, The service needs to be running inside of EPC and that thing PC needs to be associated with the private hosted zone. Now before I finish up this short lesson, let's talk about split view or split horizon DNS. Consider this scenario you have a V PC running an Amazon workspace, and to support some business applications, a private hosted zone with some records inside it. The private hosted zone is associated with VPC one on the right, meaning the workspace could use the route 53 resolver to access the private hosted zone, for example, to access the accounting records stored within the private hosted zone. Now the private hosted zone is not accessible from the public internet but what SplitView allows us to do is to create a public hosted zone with the same name, this public hosted zone might only have a subset of records that the private hosted zone has. So from the public Internet access to the public hosted zone would work in the same way as you would expect so via the ISP resolver server, Then through to the DNS root servers from that to the.org TLD servers or from there through to the animals full life name servers provided by route 53 And he records inside the public hosted zone, would be accessible, but records in the private hosted zone, which are not in the public hosted zone so accounting in this example would be inaccessible from the public Internet. And this is a common architecture, where you want to use the same domain name for public access and internal access, but with a different set of records available to each, it's something that you'll need to be comfortable with. As an architect, designing solutions, a developer integrating DNS into your applications, or an engineer, implementing this within AWS. Now that's everything I want to cover on the theory of private hosted zones. So go ahead and complete this video, and when you're ready. I look forward to you joining me in the next.